Basic Extraction
A basic extraction of the news article as well as other websites was implemented this includes the title and the main text of the website. However sites such as reuters with a bot filter make conflict with the scraper and it is unable to obtain data from it. Previous method was effective but had too much clutter but could be used as an alternative should this new method fail.